example_project/requirements/project.txt
# project.txt is a pip requirements file which describes the distributions
# required by your project to run.

--requirement=base.txt

--extra-index-url=http://dist.gondor.io/dev/
--extra-index-url=http://dist.pinaxproject.com/dev/
--extra-index-url=http://dist.pinaxproject.com/fresh-start/
--extra-index-url=http://eldarion:python4life@dist.eldarion.com/dev/

-e git+http://github.com/jlecker/django-rubberstamp.git@master#egg=rubberstamp
-e git+http://github.com/MidwestCommunications/mimesis.git@mwc-thumbnail-property#egg=mimesis
-e git+http://github.com/MidwestCommunications/mimesis-mediaman.git#egg=mediaman
-e git+https://github.com/pinax/pinax-theme-bootstrap.git#egg=pinax-theme-bootstrap

django-photologue==2.2
django-taggit==0.9.3
nashvegas==0.6.3
django_compressor==0.9.2
psycopg2==2.3
gondor==1.0

celery==2.3.3
django-celery==2.3.3
PIL==1.1.7
easy_thumbnails==1.0-alpha-18
